newsletter_optimizer/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0
pyyaml>=6.0.0
flask>=3.0.0
feedparser>=6.0.0
tiktoken>=0.5.0
pypdf>=4.0.0
